first primitive versin of classifying packages according to length of non-installability

+++ scripts/cleanup.py
@@ -43,6 +43,7 @@ def cleanup(timestamps_keep,timestamps_known,timestamp_this,scenarios):
                 shutil.rmtree(path1)
             else:
                 for directory2 in os.listdir(path1):
+                    if directory2 == 'history': continue
                     path2=path1+'/'+directory2
                     if directory2 not in timestamps_keep and os.path.isdir(path2):
                         shutil.rmtree(path2)
